+++ tests/testthat/test-ts2clm.R
@@ -52,3 +52,9 @@ test_that("clmOnly = TRUE returns only the clim data", {
   expect_equal(ncol(res), 4)
   expect_equal(nrow(res), 366)
 })
+
+test_that("robust = TRUE switches to the slower function", {
+  t_1 <- system.time(ts2clm(sst_WA, climatologyPeriod = c("1983-01-01", "2012-12-31"), robust = TRUE))
+  t_2 <- system.time(ts2clm(sst_WA, climatologyPeriod = c("1983-01-01", "2012-12-31"), robust = FALSE))
+  expect_lt(t_2[1], t_1[1])
+})